NJ 193 - Add none of the above option to property tax
